Format. Replace fixed nonce
